finished modules

+++ exercises/08_enums/enums2.rs
@@ -7,10 +7,10 @@ struct Point {
 #[derive(Debug)]
 enum Message {
     // TODO: Define the different variants used below.
-    Resize {width: u8, height: u8},
-    Move(Point),
-    Echo(String),
-    ChangeColor(u8, u8, u8),
+    Resize {width: u32, height: u32},
+    Move (Point),
+    Echo (String),
+    ChangeColor (u32, u32, u32),
     Quit,
 }
